# Page fault before the numpy benchmark