Page faults are actually stopping most modern computers from copying fast enough. This was at the heart of a recent discussion 
https://github.com/numpy/numpy/issues/11919

Before this patch:

```
Creating NumPy arrays with 10**8 int64/float64 elements:
compress_ptr.py:29: DeprecationWarning: This function is deprecated. Please call randint(0, 1000 + 1) instead
  (np.random.random_integers(0, 1000, N), "the random distribution")
  *** ctypes.memmove() *** Time for memcpy():	0.357 s	(2.09 GB/s)
```

After this patch


```
-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Creating NumPy arrays with 10**8 int64/float64 elements:
compress_ptr.py:29: DeprecationWarning: This function is deprecated. Please call randint(0, 1000 + 1) instead
  (np.random.random_integers(0, 1000, N), "the random distribution")
  *** ctypes.memmove() *** Time for memcpy():	0.110 s	(6.78 GB/s)
```