@esc 2019-07-10
This is interesting, can you explain why using `np.full` causes a page fault?