@hmaarrfk 2019-07-10
Zeros and empty does lazy allocation where the physical memory isn't actually allocated. 

I think full calls empty, and follows it up with an explicit assignment to the whole array. Sometimes I write


```
arr = no.empty(size)
arr[...] = 0
```

I think full is a shorthand for those two lines.