@hmaarrfk 2019-07-10
I also think this stuff can be applied in the other benchmarks too.