@FrancescAlted 2019-07-10
Ha, I always wondered why a copy on empty numpy arrays was so slow.  That means that if `calloc()` would be used instead of `malloc()` in `np.empty()`, then copies would go faster.  Interesting...