@hmaarrfk 2019-07-10
i think calloc is used for zeros. It seems just as slow. The kernel can also `fake` calloc allocation these days.