@hmaarrfk 2019-07-10
@FrancescAlted Here is a more complete benchmark comparing the different array allocation styles

```
-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Creating a large NumPy array with 10**8 int64 elements:
  [       0        1        2 ... 99999997 99999998 99999999]
  Time for copying array with np.copy():                    0.334 s
  Time for copying array with np.copyto and empty_like:     0.332 s
  Time for copying array with np.copyto and zeros_like:     0.330 s
  Time for copying array with np.copyto and full_like:      0.056 s
  Time for copying array with numpy assignment:             0.056 s
```