@FrancescAlted 2019-07-10
So your benchmark seems to say that `np.zeros()` behaviour is not similar to `np.full()`, right?  I have added a comment in your code.