@hmaarrfk 2019-07-10
Correct, `np.zeros` will make a call to `calloc`, but the os may decide to lazy allocate the memory.
`np.full_like` will pass through the memory allocated at allocation time, and assign it values. This makes the expensive page faulting happen before the `time` constructs.

In truth, maybe there should be some other benchmarks added, such as:
1. Allocation.
2. Faulting.
3. Compression and copy to newly allocated memory

But at this stage, it seems like the important thing is to compare the performance of `3` is that correct?