@FrancescAlted 2019-07-10
@hmaarrfk , fully agree.  We absolutely need to compare pears with pears ;-)