@hmaarrfk 2019-07-10
I just wanted to establish this kind of best case benchmark that you seemed to have "correctly" before moving on to our exact datasets. 

It seems that in under cetain circumstances, lz4 and bloclz can injest data at the rate of `memcpy`, but yield significant compression ratios. We'll have to spend time tuning the parameters.

Have you looked into pickle5? where data wouldn't have to copied to be serialized?