@FrancescAlted 2019-07-10
No, I have never tried pickle5 myself.  Adding it as another comparison point is welcome.