@FrancescAlted 2019-07-12
@hmaarrfk is this PR ready, or you want to do some additional work?