@hmaarrfk 2019-07-12
Hey Sorry for the confusion. About 10 PRs later, the one line change I made is done. 

I'm just wondering if you want to refine the top of `bench/compare-pack-ptr.py`. My point there was to show that numpy's semantics are definitely not the concern of any slowdown. Rather that it depends on how you allocate the memory. If you think that gets accross, then there isn't anything else on my side that I would like to contribute in this PR.